refactor: split module for steps
